+++ lib/APIDigester/ModuleAnalyzerNodes.cpp
@@ -1,10 +1,11 @@
-#include "llvm/ADT/STLExtras.h"
 #include "swift/AST/ASTMangler.h"
 #include "swift/Basic/Assertions.h"
 #include "swift/Basic/Defer.h"
+#include "swift/Parse/Lexer.h"
 #include "swift/Sema/IDETypeChecking.h"
-#include <swift/APIDigester/ModuleAnalyzerNodes.h>
+#include "llvm/ADT/STLExtras.h"
 #include <algorithm>
+#include <swift/APIDigester/ModuleAnalyzerNodes.h>
 
 using namespace swift;
 using namespace ide;
